requirements.txt
numpy
pillow
pillow-heif
opencv-python
ultralytics
sentence-transformers
streamlit